+++ src/gui/maingui.cpp
@@ -202,6 +202,9 @@ MainGUI::MainGUI(MainApp* mainApp)
     QAction* acAbout = new QAction("About", this);
     connect(acAbout, SIGNAL(triggered(bool)), SLOT(slotShowAbout()));
     menuHelp->addAction(acAbout);
+    QAction* acUpdates = new QAction("Check for Updates", this);
+    connect(acUpdates, SIGNAL(triggered(bool)), m_mainApp, SLOT(checkForUpdates()));
+    menuHelp->addAction(acUpdates);
     menuBar()->addMenu(menuHelp);
 
     QSettings userPrefs;
@@ -231,8 +234,8 @@ MainGUI::MainGUI(MainApp* mainApp)
     });
 
     // checks for updates
-    connect(m_mainApp, SIGNAL(checkedForUpdates(QJsonObject)),
-            SLOT(slotCheckedForUpdates(QJsonObject)));
+    connect(m_mainApp, SIGNAL(newVersionAvailable(QVariantMap)),
+            SLOT(slotCheckedForUpdates(QVariantMap)));
     if (m_mainApp->checkUpdatesAtStart()) {
         QTimer::singleShot(1000, m_mainApp, SLOT(checkForUpdates()));
     }
@@ -338,45 +341,17 @@ void MainGUI::slotSaveAll()
     }
 }
 
-void MainGUI::slotCheckedForUpdates(const QJsonObject& j)
+void MainGUI::slotCheckedForUpdates(const QVariantMap& data)
 {
-    if (j.isEmpty()) {
-        qWarning() << "failed to check for updates (connection failed)";
-        return;
+    auto url = data.value("url").toString();
+    auto vs = data.value("version").toString();
+    if (url.isEmpty()) {
+        url = QCoreApplication::organizationDomain();
     }
-
-    QDate d(2000, 1, 1);
-    for (const QString& s : j.keys()) {
-        auto _d = QDate::fromString(s, "yyyy-MM-dd");
-        if (_d.isValid() && _d > d) {
-            d = _d;
-        }
-    }
-
-    if (d == QDate(2000, 1, 1)) {
-        qWarning() << "failed to check for updates (invalid database: keys not found)";
-        return;
-    }
-
-    auto mostRecent = j.value(d.toString("yyyy-MM-dd")).toObject();
-    QStringList currVersion = QString(EVOPLEX_VERSION).split(".");
-    if (currVersion.size() != 3 || !mostRecent.contains("major") ||
-            !mostRecent.contains("minor") || !mostRecent.contains("patch")) {
-        qWarning() << "failed to check for updates (invalid database: missing fields)";
-        return;
-    }
-
-    if (mostRecent.value("major").toInt() > currVersion.at(0).toInt() ||
-        mostRecent.value("minor").toInt() > currVersion.at(1).toInt() ||
-        mostRecent.value("patch").toInt() > currVersion.at(2).toInt())
-    {
-        auto url = mostRecent.value("url").toString("https://evoplex.org");
-        auto vs = mostRecent.value("version").toString();
-        QMessageBox::information(this, "Updates",
+    QMessageBox::information(this, "Updates",
             "A new version of Evoplex has been released!<br><br>"
             "You can download this version (" + vs + ") using the link:<br>"
             "<a href='" + url + "'>" + url + "</a>");
-    }
 }
 
 void MainGUI::slotShowLog()