+++ src/gui/maingui.h
@@ -65,7 +65,7 @@ private slots:
     void slotSaveAll();
     void slotShowLog();
     void slotShowAbout();
-    void slotCheckedForUpdates(const QJsonObject& j);
+    void slotCheckedForUpdates(const QVariantMap& data);
 
 protected:
     void closeEvent(QCloseEvent* event);